ブクマに関するところをmodel層に集める
